refactor: extract FilterFolderTile from task_filter_dropdown
